Add ERC20 Service
